# herledger-app/indexer/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HerLedger Indexer API",
    "version": "1.0.0",
    "description": "Fastify REST API exposing indexed blockchain data for HerLedger."
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "paths": {
    "/v1/health": {
      "get": {
        "summary": "Indexer health check",
        "description": "Returns indexer health and database connection status.",
        "responses": {
          "200": {
            "description": "Indexer is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "undefined",
                          "enum": [null]
                        },
                        "database": {
                          "type": "undefined",
                          "enum": [null]
                        }
                      },
                      "required": ["status", "database"]
                    },
                    "error": {
                      "type": "null"
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          },
          "503": {
            "description": "Database unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          }
        }
      }
    },
    "/v1/businesses/{businessId}": {
      "get": {
        "summary": "Get business profile",
        "description": "Returns indexed business profile data by business ID.",
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            },
            "description": "Business ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Business profile details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "businessId": {
                          "type": "string"
                        },
                        "walletAddress": {
                          "type": "string"
                        },
                        "displayName": {
                          "type": "string"
                        },
                        "metadataHash": {
                          "type": "string"
                        },
                        "active": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "businessId",
                        "walletAddress",
                        "displayName",
                        "metadataHash",
                        "active",
                        "createdAt"
                      ]
                    },
                    "error": {
                      "type": "null"
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid business ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          },
          "404": {
            "description": "Business not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          }
        }
      }
    },
    "/v1/businesses/{businessId}/events": {
      "get": {
        "summary": "Get business financial events",
        "description": "Returns paginated financial events for a given business ID.",
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of financial events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "events": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        },
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "offset": {
                              "type": "number"
                            },
                            "limit": {
                              "type": "number"
                            },
                            "count": {
                              "type": "number"
                            }
                          },
                          "required": ["offset", "limit", "count"]
                        }
                      },
                      "required": ["events", "pagination"]
                    },
                    "error": {
                      "type": "null"
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          }
        }
      }
    },
    "/v1/businesses/{businessId}/attestations": {
      "get": {
        "summary": "Get business attestations",
        "description": "Returns all attestations associated with events for a business.",
        "parameters": [
          {
            "name": "businessId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of attestations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "attestations": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        }
                      },
                      "required": ["attestations"]
                    },
                    "error": {
                      "type": "null"
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          }
        }
      }
    },
    "/v1/supported-assets": {
      "get": {
        "summary": "Supported assets info",
        "description": "Read-through endpoint for supported on-chain asset information.",
        "responses": {
          "200": {
            "description": "Supported assets response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "note": {
                          "type": "string"
                        }
                      },
                      "required": ["note"]
                    },
                    "error": {
                      "type": "null"
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          }
        }
      }
    },
    "/v1/indexer/status": {
      "get": {
        "summary": "Indexer status and metrics",
        "description": "Returns indexer checkpoint sequence and sync cycle metrics.",
        "responses": {
          "200": {
            "description": "Indexer status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "stream": {
                          "type": "string"
                        },
                        "lastLedger": {
                          "type": "number"
                        },
                        "lastCycle": {
                          "type": "object",
                          "additionalProperties": true
                        }
                      },
                      "required": ["stream", "lastLedger", "lastCycle"]
                    },
                    "error": {
                      "type": "null"
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/{hash}": {
      "get": {
        "summary": "Get Stellar transaction",
        "description": "Returns indexed Stellar transaction details by transaction hash.",
        "parameters": [
          {
            "name": "hash",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 64,
              "maxLength": 64
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transaction details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "error": {
                      "type": "null"
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          },
          "400": {
            "description": "Invalid transaction hash",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          },
          "404": {
            "description": "Transaction not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/replay/{errorId}": {
      "post": {
        "summary": "Replay dead-letter error",
        "description": "Retries indexing for a previously failed event by error ID.",
        "parameters": [
          {
            "name": "errorId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Admin authorization token"
          }
        ],
        "responses": {
          "200": {
            "description": "Replay result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "errorId": {
                          "type": "string"
                        },
                        "outcome": {
                          "type": "object"
                        },
                        "retryCount": {
                          "type": "number"
                        }
                      },
                      "required": ["errorId", "outcome", "retryCount"]
                    },
                    "error": {
                      "type": "null"
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          },
          "404": {
            "description": "Dead-letter row not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          },
          "409": {
            "description": "Already resolved or max retries exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          },
          "500": {
            "description": "Replay failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": ["code", "message"]
                    }
                  },
                  "required": ["data", "error"]
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": ["code", "message"]
      },
      "HealthSuccess": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "status": {
                "type": "undefined",
                "enum": [null]
              },
              "database": {
                "type": "undefined",
                "enum": [null]
              }
            },
            "required": ["status", "database"]
          },
          "error": {
            "type": "null"
          }
        },
        "required": ["data", "error"]
      },
      "GenericErrorResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "null"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["data", "error"]
      }
    }
  }
}